Rename Room to BoardRoom
